fix: stories for badge and bullet
