+++ src/components/Badge/__stories__/Disabled.tsx
@@ -1,0 +1,8 @@
+import { Template } from './Template'
+
+export const Disabled = Template.bind({})
+
+Disabled.args = {
+  children: ['Badge'],
+  disabled: true,
+}
