+++ src/components/Badge/__stories__/Icon.tsx
@@ -1,0 +1,9 @@
+import { Template } from './Template'
+
+export const Icon = Template.bind({})
+
+Icon.args = {
+  children: ['Badge'],
+  icon: 'information-outline',
+  variant: 'primary',
+}
