+++ src/components/Badge/__stories__/Playground.tsx
@@ -1,0 +1,7 @@
+import { Template } from './Template'
+
+export const Playground = Template.bind({})
+
+Playground.args = {
+  children: ['Badge'],
+}
