+++ src/components/Badge/__stories__/Template.tsx
@@ -1,0 +1,6 @@
+import { ComponentStory } from '@storybook/react'
+import Badge from '..'
+
+export const Template: ComponentStory<typeof Badge> = ({ ...props }) => (
+  <Badge {...props} />
+)
