+++ src/components/Bullet/__stories__/Template.tsx
@@ -1,0 +1,6 @@
+import { ComponentStory } from '@storybook/react'
+import Bullet from '..'
+
+export const Template: ComponentStory<typeof Bullet> = ({ ...props }) => (
+  <Bullet {...props} />
+)
